requirements.txt
beautifulsoup4==4.11.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==4.9.1
MarkupSafe==2.1.1
pyTelegramBotAPI==4.8.0
python-dotenv==0.21.0
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.13
Werkzeug==2.2.2
